hack/backend/GRUP57/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
email-validator
pydantic
python-dotenv
google-generativeai
jinja2
openai
